+++ plugins/@vcmap-show-case/window-tester/src/WindowExampleToggleChild.vue
@@ -1,6 +1,8 @@
 <template>
   <v-container>
-    <VcsFormButton @click="toggle">toggle child</VcsFormButton>
+    <VcsFormButton v-for="num in [1, 2]" :key="num" @click.stop="toggle(num)"
+      >toggle child {{ num }}</VcsFormButton
+    >
   </v-container>
 </template>
 
@@ -19,24 +21,26 @@
       const app = inject('vcsApp');
       const parentId = attrs['window-state'].id;
       const childId = `dynamicChild-${parentId}`;
-      const childComponent = {
-        id: childId,
-        parentId,
-        state: {
-          headerTitle: 'Example dynamicChild',
-          headerIcon: 'mdi-human-child',
-        },
-        component: WindowExampleContent,
-        slot: WindowSlot.DYNAMIC_CHILD,
-      };
 
       return {
-        toggle(e) {
-          if (app.windowManager.has(childId)) {
-            app.windowManager.remove(childId);
+        toggle(num) {
+          const id = `${childId}-${num}`;
+          if (app.windowManager.has(id)) {
+            app.windowManager.remove(id);
           } else {
-            e.stopPropagation();
-            app.windowManager.add(childComponent, owner);
+            app.windowManager.add(
+              {
+                id,
+                parentId,
+                state: {
+                  headerTitle: `Example dynamicChild ${num}`,
+                  headerIcon: 'mdi-human-child',
+                },
+                component: WindowExampleContent,
+                slot: WindowSlot.DYNAMIC_CHILD,
+              },
+              owner,
+            );
           }
         },
       };